requirements.api.txt
# Slim API-only requirements — used by Dockerfile.api for Cloud Run
# Pipeline deps (playwright, moomoo, ibkr, etc.) live in requirements.pipeline.txt

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.28.0
httpx>=0.27.0

# Data + pipeline-lite
anthropic>=0.40.0
pandas>=1.3.0
numpy>=1.21.0
requests
yfinance>=0.2.0
vaderSentiment
finvizfinance
networkx>=3.2
pyvis>=0.3.2

# GCS (signal file reading in production)
google-cloud-storage>=2.16.0

# Rate limiting
slowapi>=0.1.9

# Misc
python-telegram-bot>=13.0
pyyaml
tzdata
